Use @property for model data methods.
Generally fixes the "prepare_json" pattern, and makes accessing things that are really properties of models, even though they're not stored in the database, look like just accessing their properties. Also changed a few places that were getting a model.key().name() to just get model.key_name now that that's a simple property. More room to improve that, but not in this commit.