Moves all FIRST Event datafeeds to the new style Datafeed system.
This system separates parsing from model creation, so should be much easier to test with offline tests. The new model is less forgiving of incorrect data being passed around, throwing more exceptions if you're not using it right (like string years). The new system also is a bit more organized about how parameters are passed in, always taking 'year' first, if it's a required param, for example.

Next up will be moving the other datafeeds over to this model, building their tests, then fixing the terrible Controllers and routing that is the Datafeed system.