Implement storage of alliance selection pickings
Alliance selection pickings are stored via a JSON
property in the Event model. It supports both picks
and declines, but there is no datafeed to populate
declines yet.

Picks are sourced from a Google Docs spreadsheet
maintained by CD members (<http://www.chiefdelphi.com/forums/showpost.php?p=1351639&postcount=14>).

On the frontend, only the eliminations bracket is
updated to use the proper order of teams, in the
future it could be exposed in other views.