Fix the link to the FIRST event page
Done on the github web editor, so sorry if the formatting comes out weird.

This may not work for Championship devisions, which seems to use a brand new identifier, i.e. `http://frc-events.firstinspires.org/2016/ARCHIMEDES`. Not sure I have a good enough handle on TBAs codebase to decide what to do here.

Fixes #1417 and #1224