Update FMSAPI Match Parser to Include Octofinals
This modifies the parser to check the description for the first match
returned. If it contains the word 'Octofinal', use special numbering.

We can do this because we hit/parse each endpoint independently, and the
API typically returns matches in play order, so the octofinals will
always be first. So I think this is a safe assumption

Closes #1436