Support Octofinals in FMSAPI Parsers (#1461)
* Update FMSAPI Match Parser to Include Octofinals

This modifies the parser to check the description for the first match
returned. If it contains the word 'Octofinal', use special numbering.

We can do this because we hit/parse each endpoint independently, and the
API typically returns matches in play order, so the octofinals will
always be first. So I think this is a safe assumption

Closes #1436

* Add tests for FMSAPI Match parser

* Add tests to cover FMSAPI Alliances Parser

Includes case where we get 16 alliances back

* Support octofinals in MatchDetailParser

We can't actually yet determine if we're dealing with octofinals, since
we only have match number to go by. Not sure what to do here...

* Test MatchDetailsParser