API v3 Doc Update (#1947)
* Description and redundancy fixes.

The descriptions have been updated slightly to clarify meaning and
to correct grammar.

The `district` property in the `Event` models has been replaced with
a link to the `District_List` model, which was identical.

* Required Fields Pass & District_Ranking Model Docs

All fields in *fully-documented* models now have a `required` field list
that represents those fields that may not be `null`.

`District_Ranking` model is now documented.

* Revert multi-line JSON field as out-of-spec.

The multi-line `description` field atop the swagger spec included
the newline character, making it a multi-line string. Control characters
are not allowed in JSON strings and must be escaped.