Test suite passes in dev container
Since syncing directories is hard. Now, all of our external dependencies can be installed with `pip`.

This requires us to update to BeautifulSoup4, which seems worlds better than the version we had before (which was ancient to an unknown degree).

This also disables the legacy datafeed tests because they were broken with bs4 but we don't really care about them anymore