Update team badges to use v3 API and other cleanup
My goal was to stay as close to the original TBA team badges as
possible. Because there is no single equivalent v3 API call, this now
needs to make multiple API calls and is twice as slow due to that
(~1sec).

You will need to create a `config.php` file with the X-TBA-Auth-Key for
making API requests.

The generated teamoverlays and gradients directories are now commited to
git to keep the local tree clean when testing.