Add simple numpy stubs
This will use https://github.com/machinalis/mypy-data/ and put it in the `stubs` directory (plus some additions for what we use).

We need to use the stringified version of type hints, because at runtime, numpy is not subscriptable, so thinks like `ndarray[int]` will throw exceptions

We can hope that numpy 1.20 will be released soon, which will have some native typing support.