Implement Legacy Model Pickle Serialization (#3271)
This is the followup to https://github.com/the-blue-alliance/the-blue-alliance/pull/3234

It implements the write side, as well as some other bug fixes and cleanup. Also adds tests to assert that we can do a round trip serialize + deserialize and get the original value back.

This includes some of the serialization code from the legacy runtime, which maintains its original license. I made the basic modifications to port it to py3, and left links to where to find the original to make future debugging easier. 

I also verified that I could take pickles generated by these unit tests and be able to deserialize them in the legacy app's console. After fixing imports, of course. Which we'll need to add to the py2 code anyway.