fix template rendering for round robin bracket
This was a bug/bad idea from https://github.com/the-blue-alliance/the-blue-alliance/commit/7c6bacd31b35a345e077fdd874b84cb5d3334bd2

I changed the format to be more compatible with typing, but forgot that
we had pre-computed instances of playoff advancement, which were not
compatible.

Instead, we'll go back to the original data format, even if it means a
pyre-ignore.

Tested this works for both cases

Fixes #3294