Fix nondeterminism in secret key test
This is a test that would sometimes fail, depending on how you ran it.

I'm pretty sure it has to do with some other test successfully importing
`app` and running a request, therefore setting a secret key

However, the secret key doesn't get set until we actually try to run a
request, so that may not be true always.

Instead, let's be more explicit about forcing the `before_first_request`
callback to run, and making sure it sets the secret key properly

[nodeploy]