Add way to skip CachedQueryResult for models
For example, if a DB Query just make a point query under the hood, then
there is no reason for us to double the datastore overhead by storing it
a second time. We may as well just always load the single model
regularly.

However, for the API, we want to continue caching the dict format of the
model, so leave that.