Enable CachedQueryResult writes from py3 (#3356)
We should be good to go on this.
https://github.com/the-blue-alliance/the-blue-alliance/commit/0f79f5ff1202bc1d9d45f3161334cd0ed211510e
should mean that py2 can read data written by py2, and the test suite
we've got here should be making sure that our serialization works round
trip. I can't think of any more testing we can do on this without
enabling it on prod, so here goes.