Fix proto encoding bug with multibyte characters
This is due to the harder distinction between str/bytes in py3, which
meant the py2 copy pasta'd code had this bug.

To easily see how the bug comes up (but only when there are multi-byte
charactres in the string, most commonly with accented characters)

```
>>> s = b'PrepaTec CEM&Instituto Tecnol\xc3\xb3gico y de Estudios Superiores de Monterrey'
>>> print(len(s))
73
>>> print(len(s.decode()))
72
```

When we write the actual string into the output buffer, we will ensure
that we encode it to bytes, but we weren't doing that when we were
writing the control message for the string's length. That meant we had
an off-by-one, which is what caused later-on decidings to throw