Fix local shell connecting to local datastore
There were two things that needed fixing:
 (1) the shell launcher needs to merge the local config file with the
default one to properly resolve when someone wants to use the local
datastore

 (2) we need to run the datastore emulator bound to 0.0.0.0, so we can
access it outside the dev container. This diff changes the dev container
to run the emulator in a separate tmux window (always) instead as part
of dev_appserver. This is because dev_appserver's emulator launcher
alwys hardcodes `localhost`, and we need to specify the host ourselves.