Disable response caches by default locally
The memcache version of this had a few bugs which suggested that it
never worked. And also add a new flag to configure the Cache-Control
header.

This way, now that we're going to be supporting longer timeouts, we can
disable all of these locally so that refreshing a page on the devserver
actually refreshes it.