Fix inverted techFoulCount tiebreaker for 2023
https://www.thebluealliance.com/match/2023cmptx_sf12m1 currently has the wrong
winner assigned (blue, should be red) because the tiebreaker logic is awarding
the win to the alliance with the larger `techFoulCount` (the number of tech
fouls committed by the alliance, which award points to the opposing alliance).

This was probably overlooked when copying tiebreaker logic from 2022, where
`foulPoints` (points earned from fouls committed by the opposing alliance) was
the first tiebreaker.