Improve error messages for invalid score breakdown keys
The originally-intended behavior was for this error message to include a list of
valid keys, since is_valid_score_breakdown_key() would return either True (if
given a valid key) or a list of valid keys (if given an invalid key). This
behavior was lost in the py3 migration (#4040).

Now the error message includes a list of all invalid keys, as well as a list of
all valid keys.