Add missing 2023 breakdown fields (#5348)
* Sort 2023 breakdown keys

* Add missing 2023 breakdown keys

`autoCommunity`, `teleopCommunity`, and `links` have always been present and
describe the community game pieces.

`g405Penalty` and `h111Penalty` were added as part of FMS updates for champs
(at the same time as `extraGamePieceCount`). They likely correspond to the
introduction of the "Penalties" row, e.g. at
https://frc-events.firstinspires.org/2023/ARPKY/qualifications/1, but these
fields were false for all matches at champs so it's difficult to confirm.

* Fix error message for invalid breakdown field

* Improve error messages for invalid score breakdown keys

The originally-intended behavior was for this error message to include a list of
valid keys, since is_valid_score_breakdown_key() would return either True (if
given a valid key) or a list of valid keys (if given an invalid key). This
behavior was lost in the py3 migration (#4040).

Now the error message includes a list of all invalid keys, as well as a list of
all valid keys.