Allow tba_extraRp1 and tba_extraRp2 fields for 2023 (#5395)
## Description
This allows adding `extraRp1` and `extraRp2` fields to score breakdowns through the write API. These are not used by the FRC API, so we could consider prefixing them with `tba_` as well (I brought this up as an option on Slack but we didn't reach a consensus).

## Motivation and Context
Offseasons sometimes want to add extra ranking points for various things (e.g. the additional cargo threshold at 2022cc, rainbow link bonus at 2023mirr). Supporting breakdown fields for this purpose makes it easier to submit and retrieve data about these RPs, and also allows for disambiguation if multiple extra ranking points are used.