Locally compute event teams for event with divisions (#5408)
Having to maintain the `cmp_reg_hacks` sitevar year-over-year is annoying. Let's try to formalize the handling for CMP registration in the datafeed itsef.

This diff will handle the `skip_eventteams` special case. For (D)CMP events with divisions, once the event is in-progress/completed, we default to ignoring the team registrations fetched from upstream, and instead locally compute them using `EventTeamUpdate`, which we trigger via an enqueued task.

We do need to allow `EventTeamUpdater` to make deletes while the event is in progress, so we add a parameter that allows it.

Also fill in gaps in test coverage.